requirements.txt
opencv-python
Pillow
ultralytics
qrcode[pil]
